pyproject.toml
[project]
name = "pyweastcoastbot"
version = "0.2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite",
    "alembic",
    "asyncpg",
    "dateparser",
    "fitbit>=0.3.1",
    "greenlet",
    "humanize",
    "numpy>=1.24.4",
    "kaleido",
    "pandas>=2.3.3",
    "plotly>=5.2.1",
    "py-cord==2.5.0",
    "pycoingecko>=2.2.0",
    "python-dateutil",
    "python-dotenv",
    "requests",
    "ruff",
    "sqlmodel",
    "table2ascii",
    "yfinance"
]

[project.scripts]
pyweastcoastbot = "pyWeastCoastBot.bot.main:run"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pyWeastCoastBot*"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.alembic]
script_location = "src/pyWeastCoastBot/db/migrations"
prepend_sys_path = "."

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true
